+++ mgbell/src/main/java/com/mgbell/user/model/entity/user/User.java
@@ -1,8 +1,6 @@
-package com.mgbell.user.model.entity;
+package com.mgbell.user.model.entity.user;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -31,5 +29,4 @@ public class User {
     private String phoneNumber;
     @NotNull
     private UserRole userRole;
-
 }
