+++ mgbell/src/main/java/com/mgbell/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.mgbell.user.repository;
 
-import com.mgbell.user.model.entity.User;
+import com.mgbell.user.model.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
